+++ Changelog.md
@@ -8,6 +8,7 @@ Lilu Changelog
 - Added `matchSharedCachePath` API to support dyld cache matching on macOS 12
 - Added `_kmod` hooking for kext listening to unify kext patcher logic
 - Added zlib decompression API
+- Fixed kernel patcher support on 64-bit 10.6
 
 #### v1.5.3
 - Fixed kernel patcher support on 10.7